Improved performance of PixelFinder

+++ src/test/cpp/core/MapAccessorTest.cpp
@@ -105,6 +105,6 @@ void MapAccessorTest::testGetOtherTypeData() {
 void MapAccessorTest::testGetUntypedValue() {
 	accessor->setByte(0, 17);
 
-	const double* p = (double*) accessor->getUntypedValue(0);
+	const int8_t* p = (int8_t*) accessor->getUntypedValue(0);
 	CPPUNIT_ASSERT(*p == 17);
 }
